requirements.txt
ib_insync
pandas
numpy
pyyaml
yfinance
python-telegram-bot
ta
matplotlib
scipy
